+++ wgpu-core/src/command/bundle.rs
@@ -84,17 +84,11 @@ use alloc::{
     sync::Arc,
     vec::Vec,
 };
-use core::{
-    convert::Infallible,
-    mem,
-    num::{NonZeroU32, NonZeroU64},
-    ops::Range,
-};
+use core::{convert::Infallible, mem, num::NonZeroU32, ops::Range};
 
 use arrayvec::ArrayVec;
 use thiserror::Error;
 
-use wgpu_hal::ShouldBeNonZeroExt;
 use wgt::error::{ErrorType, WebGpuError};
 
 use crate::{
@@ -589,7 +583,7 @@ impl RenderBundleEncoder {
         buffer: Arc<Buffer>,
         index_format: wgt::IndexFormat,
         offset: wgt::BufferAddress,
-        size: Option<wgt::BufferSize>,
+        size: Option<wgt::BufferAddress>,
     ) -> Result<(), PassStateError> {
         pass_base!(self, PassErrorScope::SetIndexBuffer);
         self.base.commands.push(RenderCommand::SetIndexBuffer {
@@ -606,7 +600,7 @@ impl RenderBundleEncoder {
         buffer: Arc<Buffer>,
         index_format: wgt::IndexFormat,
         offset: wgt::BufferAddress,
-        size: Option<wgt::BufferSize>,
+        size: Option<wgt::BufferAddress>,
     ) {
         if let Err(err) = self.set_index_buffer_inner(buffer, index_format, offset, size) {
             self.device
@@ -676,7 +670,7 @@ impl RenderBundleEncoder {
         slot: u32,
         buffer: Option<Arc<Buffer>>,
         offset: wgt::BufferAddress,
-        size: Option<wgt::BufferSize>,
+        size: Option<wgt::BufferAddress>,
     ) -> Result<(), PassStateError> {
         pass_base!(self, PassErrorScope::SetVertexBuffer);
         self.base.commands.push(RenderCommand::SetVertexBuffer {
@@ -693,7 +687,7 @@ impl RenderBundleEncoder {
         slot: u32,
         buffer: Option<Arc<Buffer>>,
         offset: wgt::BufferAddress,
-        size: Option<wgt::BufferSize>,
+        size: Option<wgt::BufferAddress>,
     ) {
         if let Err(err) = self.set_vertex_buffer_inner(slot, buffer, offset, size) {
             self.device
@@ -1005,7 +999,7 @@ fn set_index_buffer(
     buffer: Arc<Buffer>,
     index_format: wgt::IndexFormat,
     offset: u64,
-    size: Option<NonZeroU64>,
+    size: Option<wgt::BufferAddress>,
 ) -> Result<(), RenderBundleErrorInner> {
     buffer.check_is_valid()?;
 
@@ -1024,19 +1018,16 @@ fn set_index_buffer(
         }
         .into());
     }
-    let end = offset
-        + buffer
-            .resolve_binding_size::<NonZeroU64>(offset, size)?
-            .get();
+    let range = buffer.resolve_vertex_or_index_binding_range(offset, size)?;
 
     state
         .buffer_memory_init_actions
         .extend(buffer.initialization_status.read().create_action(
             &buffer,
-            offset..end.get(),
+            range.clone(),
             MemoryInitKind::NeedsInitializedMemory,
         ));
-    state.set_index_buffer(buffer, index_format, offset..end.get());
+    state.set_index_buffer(buffer, index_format, range);
     Ok(())
 }
 
@@ -1046,7 +1037,7 @@ fn set_vertex_buffer(
     slot: u32,
     buffer: Option<Arc<Buffer>>,
     offset: u64,
-    size: Option<NonZeroU64>,
+    size: Option<wgt::BufferAddress>,
 ) -> Result<(), RenderBundleErrorInner> {
     let max_vertex_buffers = state.device.limits.max_vertex_buffers;
     if slot >= max_vertex_buffers {
@@ -1071,17 +1062,16 @@ fn set_vertex_buffer(
         if !offset.is_multiple_of(wgt::VERTEX_ALIGNMENT) {
             return Err(RenderCommandError::UnalignedVertexBuffer { slot, offset }.into());
         }
-        let binding_size = buffer.resolve_binding_size(offset, size)?;
-        let buffer_range = offset..(offset + binding_size.get());
+        let range = buffer.resolve_vertex_or_index_binding_range(offset, size)?;
 
         state
             .buffer_memory_init_actions
             .extend(buffer.initialization_status.read().create_action(
                 &buffer,
-                buffer_range.clone(),
+                range.clone(),
                 MemoryInitKind::NeedsInitializedMemory,
             ));
-        state.vertex.set_buffer(slot as usize, buffer, buffer_range);
+        state.vertex.set_buffer(slot as usize, buffer, range);
         if let Some(pipeline) = state.pipeline.as_deref() {
             state.vertex.update_limits(&pipeline.vertex_steps);
         }
@@ -1095,14 +1085,17 @@ fn set_vertex_buffer(
             )
             .into());
         }
-        if let Some(size) = size {
-            return Err(RenderCommandError::from(
-                crate::binding_model::BindingError::UnbindingVertexBufferSizeNotZero {
-                    slot,
-                    size: size.get(),
-                },
-            )
-            .into());
+        match size {
+            Some(size) if size != 0 => {
+                return Err(RenderCommandError::from(
+                    crate::binding_model::BindingError::UnbindingVertexBufferSizeNotZero {
+                        slot,
+                        size,
+                    },
+                )
+                .into());
+            }
+            _ => {}
         }
 
         state.vertex.clear_buffer(slot as usize);
@@ -1729,7 +1722,7 @@ impl IndexState {
                 buffer: self.buffer.clone(),
                 index_format: self.format,
                 offset: self.range.start,
-                size: NonZeroU64::new(binding_size),
+                size: Some(binding_size),
             })
         } else {
             None