+++ wgpu/src/api/render_bundle_encoder.rs
@@ -97,8 +97,7 @@ impl<'a> RenderBundleEncoder<'a> {
             &buffer_slice.buffer.inner,
             index_format,
             buffer_slice.offset,
-            // TODO(https://github.com/gfx-rs/wgpu/issues/3170): Empty slices should be supported here
-            Some(buffer_slice.size_expect_nonzero()),
+            Some(buffer_slice.size),
         );
     }
 
@@ -126,8 +125,7 @@ impl<'a> RenderBundleEncoder<'a> {
                 slot,
                 Some(&buffer_slice.buffer.inner),
                 buffer_slice.offset,
-                // TODO(https://github.com/gfx-rs/wgpu/issues/3170): Empty slices should be supported here
-                Some(buffer_slice.size_expect_nonzero()),
+                Some(buffer_slice.size),
             );
         } else {
             self.inner.set_vertex_buffer(slot, None, 0, None);